challenge2/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
nltk==3.9.1
jinja2==3.1.6
python-dotenv==1.1.0
pydantic==2.11.1
google-generativeai==0.8.4
